pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tradingtest"
version = "0.1.0"
description = "Unified Polymarket trading system with multi-strategy arena"
requires-python = ">=3.10"
# #48: Synced with requirements.txt - removed unused deps
# (jinja2, python-telegram-bot, requests, websocket-client, apscheduler,
#  plotly, pydantic, scikit-learn, pandas)
dependencies = [
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
    "xgboost>=2.0.0",
    "numpy>=1.24.0",
    "rapidfuzz>=3.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
]

[project.optional-dependencies]
scraper = ["playwright>=1.40.0"]
rust = ["maturin>=1.3.0"]
all = ["playwright>=1.40.0", "maturin>=1.3.0"]

[project.scripts]
tradingtest = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
